+++ makepak/src/tomlinput.rs
@@ -48,7 +48,7 @@ pub struct AnimImgDepInput {
 #[derive(Debug, Deserialize)]
 pub struct CharaTemplateDepInput {
     pub race: gamedata::chara::Race,
-    pub gen_weight: u32,
+    pub gen_weight: f32,
     pub gen_level: u32,
     pub max_hp: i32,
     pub str: u32,
